requirements.txt
joblib==1.2.0
numpy==1.23.5
Pillow==9.3.0
Pillow==10.0.0
streamlit==1.15.1
